Isolated fragment tests
